detectors/image_detector/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
opencv-python
numpy
Pillow
python-multipart